moved TroubleshootingModalComponent.tsx to shared
